refactor(css): move export stringification into css module generator
